refactor: extract namespace fetching logic into a common hook
